openraft

Module docs

Openraft Document

If you’re starting to build an application with Openraft, check out

getting_started.

faq,

To maintain an Openraft cluster, e.g., add or remove nodes, refer to

cluster_control :

cluster_formation describes how to form a cluster;

dynamic membership describes how to add or remove nodes without downtime;

node lifecycle describes the transition of a node’s state;

When upgrading an Openraft application, consult:

upgrade_guide :

v0.6-to-v0.7;

v0.7-to-v0.8;

v0.8.3-to-v0.8.4;

v0.8-to-v0.9;

To learn about the data structures used in Openraft and the commit protocol, see

feature_flags;

data :

Vote is the core in a distributed system;

Log pointers shows how Openraft tracks entries in the log;

Leader lease explains leader lease for a Leader and a Follower;

Extended membership explains how members are organized in Openraft;

Effective membership explains when membership config takes effect;

components explains the components in Openraft;

RaftStateMachine is the core API for managing the state machine and snapshot functionalities;

protocol :

replication;

leader_lease;

log_replication;

snapshot_replication;

Contributors who want to understand the internals of Openraft can find relevant information in

internal :

Architecture shows the overall architecture of Openraft;

Threading describes the threading model of Openraft;

Finally, the archived and discarded documents:

obsolete describes obsolete design documents and why they are discarded;

blank-log-heartbeeat;

fast-commit;

Modules

cluster_control

Managing Clusters: Adding/Removing Nodes

components

Components and sub systems of the openraft project.

data

Data structures used by the Openraft protocol, such log, vote, snapshot, membership etc.

faq

FAQ

feature_flags

Feature flags

getting_started

Getting Started with Openraft

internal

Openraft internal design and implementation

obsolete

Obsolete Designs

protocol

The protocol used by Openraft to replicate data.

upgrade_guide

Guide for upgrading from older version Openraft