openraft::docs

Module upgrade_guide

Guide for upgrading from older version Openraft

When upgrading to a new version: First, check the change log for the version to upgrade to. A commit message starting with these keywords needs attention:

DataChange: introduces breaking changes to data types. A data migration tool or a compatible layer is required for the upgrade. See below.

Change: introduces breaking changes. Your application needs adjustment to pass compilation. If storage related data structure changed too, a data migration tool is required for the upgrade. See below.

Feature: introduces non-breaking new features. Your application should compile without modification.

Improve: No breaking new features. Your application should compile without modification.

Fix: bug fix. No modification is required.

Upgrade from v0.8 to v0.9:

Change log v0.9.0

Guide for upgrading v0.8 to v0.9

Upgrade from v0.8.3 to v0.8.4:

Change log v0.8.4

Guide for upgrading v0.8.3 to v0.8.4

Upgrade from v0.7 to v0.8:

Change log v0.8

Guide for upgrading v0.7 to v0.8

Upgrade from v0.6.8 to v0.7.0:

Change log v0.7.0

Guide for upgrading v0.6 to v0.7

Upgrade from v0.6.5 to v0.6.6:

Change log v0.6.6

just modify application code to pass compile.

API changes: struct fields changed in StorageIOError and Violation.

Data changes: none.

Modules

upgrade_06_07

See: Guide for upgrading v0.6 to v0.7

upgrade_07_08

See: Guide for upgrading v0.7 to v0.8

upgrade_08_09

Guide for upgrading from v0.8 to v0.9:

upgrade_083_084

See: Guide for upgrading v0.8.3 to v0.8.4