openraft

Module error

Error types exposed by this crate.

Modules

decompose

into_ok

Structs

EmptyMembership

ForwardToLeader

InProgress

LackEntry

LearnerNotFound

NetworkError

Error that indicates a temporary network error and when it is returned, Openraft will retry immediately.

NotAllowed

NotInMembers

PayloadTooLarge

Error indicating that an RPC is too large and cannot be sent.

QuorumNotEnough

RemoteError

ReplicationClosed

Replication is closed intentionally.

SnapshotMismatch

Timeout

Unreachable

Error indicating a node is unreachable. Retries should be delayed.

Enums

ChangeMembershipError

The set of errors which may take place when requesting to propose a config change.

CheckIsLeaderError

An error related to a is_leader request.

ClientWriteError

An error related to a client write request.

Fatal

Fatal is unrecoverable and shuts down raft at once.

Infallible

InitializeError

The set of errors which may take place when initializing a pristine Raft node.

InstallSnapshotError

NoForward

A place holder to mark RaftError won’t have a ForwardToLeader variant.

RPCError

Error occurs when invoking a remote raft API.

RaftError

RaftError is returned by API methods of Raft.

StreamingError

Error occurred when streaming local data to a remote raft node.