openraft::error

Enum RaftError

pub enum RaftError<NID, E = Infallible>
where
    NID: NodeId,
{
    APIError(E),
    Fatal(Fatal<NID>),
}

RaftError is returned by API methods of Raft.

It is either a Fatal error indicating that Raft is no longer running, such as underlying IO error, or an API error E.

Variants

APIError(E)

Fatal(Fatal<NID>)

Implementations

impl<NID, E> RaftError<NID, E>
where NID: NodeId, E: Debug,

pub fn api_error(&self) -> Option<&E>

Return a reference to Self::APIError.

pub fn into_api_error(self) -> Option<E>

Try to convert self to APIError.

pub fn fatal(&self) -> Option<&Fatal<NID>>

Return a reference to Self::Fatal.

pub fn into_fatal(self) -> Option<Fatal<NID>>

Try to convert self to Fatal error.

pub fn forward_to_leader<N>(&self) -> Option<&ForwardToLeader<NID, N>>
where N: Node, E: TryAsRef<ForwardToLeader<NID, N>>,

Return a reference to ForwardToLeader if Self::APIError contains it.

pub fn into_forward_to_leader<N>(self) -> Option<ForwardToLeader<NID, N>>
where N: Node, E: TryInto<ForwardToLeader<NID, N>>,

Try to convert self to ForwardToLeader error if APIError is a ForwardToLeader error.