openraft

Module impls

Collection of implementations of usually used traits defined by Openraft

Re-exports

pub use crate::async_runtime::TokioRuntime;

pub use crate::entry::Entry;

Structs

BasicNode

An implementation of trait Node that contains minimal node information.

EmptyNode

EmptyNode is an implementation of trait Node that contains nothing.

OneshotResponder

A Responder implementation that sends the response via a oneshot channel.