openraft

Module log_id

This mod defines the identity of a raft log and provides supporting utilities to work with log id related types.

Structs

LogId

The identity of a raft log.

Traits

LogIdOptionExt

This helper trait extracts information from an Option<LogId>.

LogIndexOptionExt

This helper trait extracts information from an Option<LogIndex>.

RaftLogId

Defines API to operate an object that contains a log-id, such as a log entry or a log id.