openraft

Module raft

Public Raft interface and data types.

Raft serves as the primary interface to a Raft node, facilitating all interactions with the underlying RaftCore.

While RaftCore operates as a singleton within an application, Raft instances are designed to be cheaply cloneable. This allows multiple components within the application that require interaction with RaftCore to efficiently share access.

Re-exports

pub use crate::RaftTypeConfig;

Modules

responder

API to consumer a response when a client write request is completed.

trigger

Trigger an action to RaftCore by external caller.

Structs

AppendEntriesRequest

An RPC sent by a cluster leader to replicate log entries (5.3), and as a heartbeat (5.2).

ClientWriteResponse

The response to a client-request.

InstallSnapshotRequest

An RPC sent by the Raft leader to send chunks of a snapshot to a follower (7).

InstallSnapshotResponse

The response to an InstallSnapshotRequest.

Raft

The Raft API.

RuntimeConfigHandle

RuntimeConfigHandle is an interface to update runtime config.

SnapshotResponse

The response to Raft::install_full_snapshot API.

VoteRequest

An RPC sent by candidates to gather votes (5.2).

VoteResponse

The response to a VoteRequest.

Enums

AppendEntriesResponse

The response to an AppendEntriesRequest.

Type Aliases

ClientWriteResult

The result of a write request to Raft.