openraft

Module storage

The Raft storage interface and data types.

Structs

Adaptor

An adapter that allows an implementation of RaftStorage to be used in the latest framework.

LogApplied

A oneshot callback for completion of applying logs to state machine.

LogFlushed

A oneshot callback for completion of log io operation.

LogState

The state about logs.

Snapshot

The data associated with the current snapshot.

SnapshotMeta

The metadata of a snapshot.

SnapshotSignature

A small piece of information for identifying a snapshot and error tracing.

StorageHelper

StorageHelper provides additional methods to access a RaftLogStorage and RaftStateMachine implementation.

Traits

RaftLogReader

A trait defining the interface for a Raft log subsystem.

RaftLogReaderExt

RaftLogStorage

API for log store.

RaftLogStorageExt

Extension trait for RaftLogStorage to provide utility methods.

RaftSnapshotBuilder

A trait defining the interface for a Raft state machine snapshot subsystem.

RaftStateMachine

API for state machine and snapshot.

RaftStorage

A trait defining the interface for a Raft storage system.