openraft::storage

Struct Adaptor

pub struct Adaptor<C, S>
where
    C: RaftTypeConfig,
    S: RaftStorage<C>,
{ /* private fields */ }

An adapter that allows an implementation of RaftStorage to be used in the latest framework.

It hide an implementation of RaftStorage behind a RWLock. Therefore, it provides full functionalities but without any parallelism.

Adaptor implements both RaftLogStorage and RaftStateMachine, and just pass the calls to the underlying RaftStorage.

To use the old RaftStorage implementation in the latest framework:

let store = MyRaftStorage::new();
let (log_store, state_machine) = Adaptor::new(store);
Raft::new(1, Arc::new(Config::default()), MyNetwork::default(), log_store, state_machine);

Implementations

impl<C, S> Adaptor<C, S>
where C: RaftTypeConfig, S: RaftStorage<C>,

pub fn new(store: S) -> (Self, Self)

Create a RaftLogStorage and a RaftStateMachine upon an implementation of RaftStorage.

pub async fn storage_mut(&self) -> RwLockWriteGuard<'_, S>

Get a write lock of the underlying storage.

pub async fn storage(&self) -> RwLockReadGuard<'_, S>

Get a read lock of the underlying storage.