openraft::storage

Trait RaftLogReaderExt

pub trait RaftLogReaderExt<C>: RaftLogReader<C> + Send
where
    C: RaftTypeConfig,
{
    // Provided methods
    fn try_get_log_entry(
        &mut self,
        log_index: u64,
    ) -> impl Future<Output = Result<Option<C::Entry>, StorageError<C::NodeId>>> + Send { ... }
    fn get_log_entries<RB: RangeBounds<u64> + Clone + Debug + OptionalSend + OptionalSync>(
        &mut self,
        range: RB,
    ) -> impl Future<Output = Result<Vec<C::Entry>, StorageError<C::NodeId>>> + Send { ... }
    fn get_log_id(
        &mut self,
        log_index: u64,
    ) -> impl Future<Output = Result<LogId<C::NodeId>, StorageError<C::NodeId>>> + Send { ... }
}

Provided Methods

fn try_get_log_entry( &mut self, log_index: u64, ) -> impl Future<Output = Result<Option<C::Entry>, StorageError<C::NodeId>>> + Send

Try to get an log entry.

It does not return an error if the log entry at log_index is not found.

fn get_log_entries<RB: RangeBounds<u64> + Clone + Debug + OptionalSend + OptionalSync>( &mut self, range: RB, ) -> impl Future<Output = Result<Vec<C::Entry>, StorageError<C::NodeId>>> + Send

Get a series of log entries from storage.

Similar to try_get_log_entries except an error will be returned if there is an entry not found in the specified range.

fn get_log_id( &mut self, log_index: u64, ) -> impl Future<Output = Result<LogId<C::NodeId>, StorageError<C::NodeId>>> + Send

Get the log id of the entry at index.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.