openraft::type_config

Trait TypeConfigExt

pub trait TypeConfigExt: RaftTypeConfig {
    // Provided methods
    fn now(    ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Instant { ... }
    fn sleep(
        duration: Duration,
    ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Sleep { ... }
    fn sleep_until(
        deadline: <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Instant,
    ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Sleep { ... }
    fn timeout<R, F: Future<Output = R> + OptionalSend>(
        duration: Duration,
        future: F,
    ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Timeout<R, F> { ... }
    fn timeout_at<R, F: Future<Output = R> + OptionalSend>(
        deadline: <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Instant,
        future: F,
    ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Timeout<R, F> { ... }
    fn oneshot<T>(    ) -> (<<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::OneshotSender<T>, <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::OneshotReceiver<T>)
       where T: OptionalSend { ... }
    fn spawn<T>(
        future: T,
    ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::JoinHandle<T::Output>
       where T: Future + OptionalSend + 'static,
             T::Output: OptionalSend + 'static { ... }
}

Collection of utility methods to RaftTypeConfig function.

Provided Methods

fn now() -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Instant

Returns the current time.

fn sleep( duration: Duration, ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Sleep

Wait until duration has elapsed.

fn sleep_until( deadline: <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Instant, ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Sleep

Wait until deadline is reached.

fn timeout<R, F: Future<Output = R> + OptionalSend>( duration: Duration, future: F, ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Timeout<R, F>

Require a Future to complete before the specified duration has elapsed.

fn timeout_at<R, F: Future<Output = R> + OptionalSend>( deadline: <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Instant, future: F, ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::Timeout<R, F>

Require a Future to complete before the specified instant in time.

fn oneshot<T>() -> (<<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::OneshotSender<T>, <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::OneshotReceiver<T>)
where T: OptionalSend,

Creates a new one-shot channel for sending single values.

This is just a wrapper of AsyncRuntime::oneshot.

fn spawn<T>( future: T, ) -> <<Self as RaftTypeConfig>::AsyncRuntime as AsyncRuntime>::JoinHandle<T::Output>
where T: Future + OptionalSend + 'static, T::Output: OptionalSend + 'static,

Spawn a new task.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.